Change "amount" type of Activity class to double, to take into account km registered by bike.

See merge request cse1105/2018-2019/oopp-group-66/template!106

+++ src/main/java/app/controllers/RouteController.java
@@ -145,7 +145,7 @@ public class RouteController {
 
         for (Activity a : activities) {
 
-            long amount = a.getAmount();
+            double amount = a.getAmount();
             double xp_points = a.getXp_points();
             String category = "";
             switch ((int)a.getCategory_id()) {
